--story=856902347 op编辑扩展点无效

+++ src/backend/ci/ext/tencent/store/api-store-service/src/main/kotlin/com/tencent/devops/store/pojo/UpdateSettingInfo.kt
@@ -13,5 +13,5 @@ data class UpdateSettingInfo(
     @ApiModelProperty("是否官方认证", required = true)
     val certificationFlag: Boolean?,
     @ApiModelProperty("权重", required = true)
-    val weight: Float?
+    val weight: Int?
 )
